Updated function logout

+++ views/users/users_list.php
@@ -2,39 +2,53 @@
 if (session_status() == PHP_SESSION_NONE) {
     session_start();
 }
- if (isset($_SESSION['user_id'])) : ?>
+
+if (isset($_SESSION['user_id'])) :
+    $userRole = $_SESSION['user_role']; // Get the user role from session
+?>
 <div class="container mt-4">
-<h1>Users List</h1>
-<a href="/users/create" class="btn btn-primary">Create</a>
-<table class="table">
-<thead>
+    <h1>Users List</h1>
+
+    <!-- Display "Create" button for admin users only -->
+    <?php if ($userRole == 'admin') : ?>
+        <a href="/users/create" class="btn btn-primary">Create</a>
+    <?php endif; ?>
+
+    <table class="table">
+        <thead>
             <tr>
                 <th>Id</th>
                 <th>Username</th>
                 <th>Email</th>
                 <th>Role</th>
-                <th>Actions</th>
+                <!-- Display "Actions" column only for admin users -->
+                <?php if ($userRole == 'admin') : ?>
+                    <th>Actions</th>
+                <?php endif; ?>
             </tr>
         </thead>
         <tbody>
-            <?php foreach ($users as $index=> $user): ?>
+            <?php foreach ($users as $user): ?>
             <tr>
-                <td><?= $index+1 ?></td>
+                <td><?= $user['id'] ?></td>
                 <td><?= $user['name'] ?></td>
                 <td><?= $user['email'] ?></td>
                 <td><?= $user['role'] ?></td>
-                <td>
-                    <a href="/users/edit/<?= $user['id'] ?>"><i class="material-icons">edit</i></a>
-                    <a href="/users/delete/<?= $user['id'] ?>"><i class="material-icons text-danger">delete</i></a>
-                </td>
+                
+                <!-- Display action buttons only for admin users -->
+                <?php if ($userRole == 'admin') : ?>
+                    <td>
+                        <a href="/users/edit/<?= $user['id'] ?>"><i class="material-icons">edit</i></a>
+                        <a href="/users/delete/<?= $user['id'] ?>"><i class="material-icons text-danger">delete</i></a>
+                    </td>
+                <?php endif; ?>
             </tr>
-        <?php endforeach; ?>
+            <?php endforeach; ?>
         </tbody>
-</table>
-   
-</div>  
+    </table>
+</div>
 <?php 
 else: 
     $this->redirect("/users/login"); 
-endif;   
+endif; 
 ?>
